fix-flaky: MultiCodecReindexTests (#10980)
Segement merges during source index indexing is leading to flakiness.
Let's stop merges altogether.

(cherry-picked from commit 0d3d7aed741db3601d0a3fee2104f51f6c7ba8ad)
